Add a ?blueprint-url Query API option to enable sourcing Blueprints from external URLs
