+++ packages/docs/site/docs/08-query-api/01-index.md
@@ -25,6 +25,7 @@ You can go ahead and try it out. The Playground will automatically install the t
 | ---------------------- | ------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | `php`                  | `8.0`         | Loads the specified PHP version. Supported values: `5.6`, `7.0`, `7.1`, `7.2`, `7.3`, `7.4`, `8.0`, `8.1`, `8.2`, `latest`                                                                                                                                                                                                                                     |
 | `wp`                   | `latest`      | Loads the specified WordPress version. Supported values: `5.9`, `6.0`, `6.1`, `6.2`, `6.3`, `latest`, `nightly`, `beta`                                                                                                                                                                                                                                        |
+| `blueprint-url`        |               | The URL of the Blueprint that will be used to configure this Playground instance.                                                                                                                                                                                                                                                                              |
 | `php-extension-bundle` |               | Loads a bundle of PHP extensions. Supported bundles: `kitchen-sink` (for gd, mbstring, iconv, libxml, xml, dom, simplexml, xmlreader, xmlwriter)                                                                                                                                                                                                               |
 | `networking`           | `yes` or `no` | Enables or disables the networking support for Playground. Defaults to `yes`                                                                                                                                                                                                                                                                                   |
 | `plugin`               |               | Installs the specified plugin. Use the plugin name from the plugins directory URL, e.g. for a URL like `https://wordpress.org/plugins/wp-lazy-loading/`, the plugin name would be `wp-lazy-loading`. You can pre-install multiple plugins by saying `plugin=coblocks&plugin=wp-lazy-loading&…`. Installing a plugin automatically logs the user in as an admin |
