+++ packages/playground/website/src/lib/resolve-blueprint.ts
@@ -1,0 +1,67 @@
+import { Blueprint } from '@wp-playground/client';
+import { makeBlueprint } from './make-blueprint';
+
+const query = new URL(document.location.href).searchParams;
+const fragment = decodeURI(document.location.hash || '#').substring(1);
+
+export async function resolveBlueprint() {
+	let blueprint: Blueprint;
+	/*
+	 * Support passing blueprints via query parameter, e.g.:
+	 * ?blueprint-url=https://example.com/blueprint.json
+	 */
+	if (query.has('blueprint-url')) {
+		const url = query.get('blueprint-url');
+		const response = await fetch(url!, {
+			credentials: 'omit',
+		});
+		blueprint = await response.json();
+	} else if (fragment.length) {
+		/*
+		 * Support passing blueprints in the URI fragment, e.g.:
+		 * /#{"landingPage": "/?p=4"}
+		 */
+		try {
+			blueprint = JSON.parse(fragment);
+			// Allow overriding the preferred versions using query params
+			// generated by the version switchers.
+			if (query.get('php') || query.get('wp')) {
+				if (!blueprint.preferredVersions) {
+					blueprint.preferredVersions = {} as any;
+				}
+				blueprint.preferredVersions!.php =
+					(query.get('php') as any) ||
+					blueprint.preferredVersions!.php ||
+					'8.0';
+				blueprint.preferredVersions!.wp =
+					query.get('wp') ||
+					blueprint.preferredVersions!.wp ||
+					'latest';
+			}
+		} catch (e) {
+			// Noop
+		}
+	}
+
+	// If no blueprint was passed, prepare one based on the query params.
+	// @ts-ignore
+	if (typeof blueprint === 'undefined') {
+		const features: Blueprint['features'] = {};
+		// Networking is enabled by default, so we only need to disable it
+		// if the query param is explicitly set to "no".
+		if (query.get('networking') === 'no') {
+			features['networking'] = false;
+		}
+		blueprint = makeBlueprint({
+			php: query.get('php') || '8.0',
+			wp: query.get('wp') || 'latest',
+			theme: query.get('theme') || undefined,
+			features,
+			plugins: query.getAll('plugin'),
+			landingPage: query.get('url') || undefined,
+			phpExtensionBundles: query.getAll('php-extension-bundle') || [],
+		});
+	}
+
+	return blueprint;
+}
