+++ packages/playground/website/src/main.tsx
@@ -6,8 +6,6 @@ import css from './style.module.css';
 import buttonCss from './components/button/style.module.css';
 import './styles.css';
 
-import { makeBlueprint } from './lib/make-blueprint';
-import type { Blueprint } from '@wp-playground/blueprints';
 import PlaygroundConfigurationGroup from './components/playground-configuration-group';
 import { PlaygroundConfiguration } from './components/playground-configuration-group/form';
 import { SupportedPHPVersions } from '@php-wasm/universal';
@@ -15,47 +13,10 @@ import { StorageType, StorageTypes } from './types';
 import { ResetSiteMenuItem } from './components/toolbar-buttons/reset-site';
 import { DownloadAsZipMenuItem } from './components/toolbar-buttons/download-as-zip';
 import { RestoreFromZipMenuItem } from './components/toolbar-buttons/restore-from-zip';
+import { resolveBlueprint } from './lib/resolve-blueprint';
 
 const query = new URL(document.location.href).searchParams;
-
-/*
- * Support passing blueprints in the URI fragment, e.g.:
- * /#{"landingPage": "/?p=4"}
- */
-const fragment = decodeURI(document.location.hash || '#').substring(1);
-let blueprint: Blueprint;
-try {
-	blueprint = JSON.parse(fragment);
-	// Allow overriding the preferred versions using query params
-	// generated by the version switchers.
-	if (query.get('php') || query.get('wp')) {
-		if (!blueprint.preferredVersions) {
-			blueprint.preferredVersions = {} as any;
-		}
-		blueprint.preferredVersions!.php =
-			(query.get('php') as any) ||
-			blueprint.preferredVersions!.php ||
-			'8.0';
-		blueprint.preferredVersions!.wp =
-			query.get('wp') || blueprint.preferredVersions!.wp || 'latest';
-	}
-} catch (e) {
-	const features: Blueprint['features'] = {};
-	// Networking is enabled by default, so we only need to disable it
-	// if the query param is explicitly set to "no".
-	if (query.get('networking') === 'no') {
-		features['networking'] = false;
-	}
-	blueprint = makeBlueprint({
-		php: query.get('php') || '8.0',
-		wp: query.get('wp') || 'latest',
-		theme: query.get('theme') || undefined,
-		features,
-		plugins: query.getAll('plugin'),
-		landingPage: query.get('url') || undefined,
-		phpExtensionBundles: query.getAll('php-extension-bundle') || [],
-	});
-}
+const blueprint = await resolveBlueprint();
 
 // @ts-ignore
 const opfsSupported = typeof navigator?.storage?.getDirectory !== 'undefined';
